@@ -26,12 +26,16 @@
 		{
 			"name": "Thomas Mulhall",
 			"role": "Developer"
+		},
+		{
+			"name": "Alexander Gesinn",
+			"role": "Developer"
 		}
 	],
 	"require": {
-		"php": ">=5.5",
+		"php": ">=5.6",
 		"composer/installers": "1.*,>=1.0.1",
-		"mediawiki/semantic-media-wiki": "~2.3|~3.0",
+		"mediawiki/semantic-media-wiki": "~2.5|~3.0",
 		"mediawiki/page-forms": ">=4.0.2"
 	},
 	"require-dev": {
