@@ -1,75 +1,47 @@
 <?php
 
-namespace SFS;
-
-use ResourceLoader;
-use ParserOutput;
-
 /**
  * @license GNU GPL v2+
  * @since 1.3
  *
  * @author mwjames
+ * @author Alexander Gesinn
  */
+
+namespace SFS;
+
+//use MWDebug;
+
 class Output {
 
 	/**
 	 * @var array
 	 */
-	private static $headItems = array();
+	private static $headItems = [];
 
 	/**
-	 * @var array
-	 */
-	private static $resourceModules = array();
-
-	/**
-	 * @since 1.3
+	 * Add an array of SF_Select field parameters as defined in Page Form's field tag.
 	 *
-	 * @param string $moduleName
+	 * This will later be added to $wgOut so that JS can access it via mw.config.get
+	 *
+	 * @param array $data
 	 */
-	public static function addModule( $moduleName ) {
-		self::$resourceModules[$moduleName] = $moduleName;
+	public static function addToHeadItem( Array $data = [] ) {
+		return self::$headItems[] = $data;
 	}
 
 	/**
-	 * @since 1.3
+	 * Commit all SF_Select field parameters to Output
 	 *
-	 * @param string $id
-	 * @param string $data
 	 */
-	public static function addToHeadItem( $id, $data = '' ) {
-		return self::$headItems[$id] = self::makeVariablesScript( array( $id => json_encode( $data ) ) );
+	public static function commitToParserOutput() {
+		global $wgOut;	# is there a better way to get $output/$parser without using a global? (testability!)
+
+		// to be used in JS like:
+		// var SFSelect_fobjs = $.parseJSON( mw.config.get( 'sf_select' ) );
+		$wgOut->addJsConfigVars('sf_select', json_encode( self::$headItems ));
+
+		//self::$resourceModules = array();
+		//self::$headItems = array();
 	}
-
-	/**
-	 * @since 1.3
-	 *
-	 * @param ParserOutput $parserOutput
-	 */
-	public static function commitToParserOutput( ParserOutput $parserOutput ) {
-
-		foreach ( self::$headItems as $key => $item ) {
-			$parserOutput->addHeadItem( "\t\t" . $item . "\n", $key );
-		}
-
-		$parserOutput->addModules( array_values( self::$resourceModules ) );
-
-		self::$resourceModules = array();
-		self::$headItems = array();
-	}
-
-	private static function makeVariablesScript( $data ) {
-
-		if ( $data ) {
-			return \Html::inlineScript(
-				ResourceLoader::makeLoaderConditionalScript(
-					ResourceLoader::makeConfigSetScript( $data )
-				)
-			);
-		}
-
-		return '';
-	}
-
 }
