@@ -1,13 +1,5 @@
 <?php
 
-namespace SFS;
-
-use ApiBase;
-use Parser;
-use ParserOptions;
-use ParserOutput;
-use Title;
-
 /**
  * API modules to communicate with the back-end
  *
@@ -17,6 +9,15 @@ use Title;
  * @author Jason Zhang
  * @author Toni Hermoso Pulido
  */
+
+namespace SFS;
+
+use ApiBase;
+use Parser;
+use ParserOptions;
+use ParserOutput;
+use Title;
+
 class ApiSemanticFormsSelect extends ApiBase {
 
 	/**
@@ -29,7 +30,7 @@ class ApiSemanticFormsSelect extends ApiBase {
 		$parser->mOptions = new ParserOptions();
 		$parser->mOutput = new ParserOutput();
 
-		$apiRequestProcessor = new ApiRequestProcessor( $parser );
+		$apiRequestProcessor = new \SFS\ApiSemanticFormsSelectRequestProcessor( $parser );
 		$apiRequestProcessor->setDebugFlag( $GLOBALS['wgSF_Select_debug'] );
 
 		$resultValues = $apiRequestProcessor->getJsonDecodedResultValuesForRequestParameters(
@@ -48,39 +49,39 @@ class ApiSemanticFormsSelect extends ApiBase {
 	 * @see ApiBase::getAllowedParams
 	 */
 	public function getAllowedParams() {
-		return array(
-			'approach' => array(
+		return [
+			'approach' => [
 				ApiBase::PARAM_TYPE => 'string',
 				ApiBase::PARAM_REQUIRED => true
-			),
-			'query' => array(
+			],
+			'query' => [
 				ApiBase::PARAM_TYPE => 'string',
 				ApiBase::PARAM_REQUIRED => true
-			),
-			'sep' => array(
+			],
+			'sep' => [
 				ApiBase::PARAM_TYPE => 'string',
 				ApiBase::PARAM_REQUIRED => false
-			)
-		);
+			]
+		];
 	}
 
 	/**
 	 * @see ApiBase::getDescription
 	 */
 	public function getDescription() {
-		return array(
+		return [
 			'API for providing SemanticFormsSelect values'
-		);
+		];
 	}
 
 	/**
 	 * @see ApiBase::getParamDescription
 	 */
 	public function getParamDescription() {
-		return array(
+		return [
 			'approach' => 'The actual approach: function or smw',
 			'query' => 'The query of the former'
-		);
+		];
 	}
 
 	/**
